requirements_nvidia.txt
# Requires python 3.11
albumentations==2.0.8
h5py==3.14.0
matplotlib==3.9.4
pandas==2.3.1
pipdeptree==2.28.0
pytest-html==4.1.1
rich==14.0.0
setuptools==78.1.1
wheel==0.45.1
pyaml==25.7.0
pydicom==2.4.4
tqdm==4.67.1
python-dotenv==1.1.1
scikit_learn==1.6.1
omegaconf==2.3.0
hydra-core==1.3.2
natsort==8.4.0
lightning==2.5.2
torchdata==0.11.0
torchsummary==1.5.1
torchvision==0.22.1
